Publish to Central
Updated Gradle and with it all other dependencies. Minimum required
Kotlin version is now 1.3, anything older is anyways deprecated so this
should not be an issue for anyone. I also updated the publishing logic
from Bintray to Central since JFrog is going to sunset Bintray. Wanted
to do this from the start, only the right incentive was missing.